requirements.txt
Flask==3.0.2
python-dateutil==2.8.2
Werkzeug==3.0.1
Jinja2==3.1.3 